python_scipts/ski-processor/requirements.txt
Flask==2.3.3
requests==2.31.0
google-cloud-storage==2.10.0
functions-framework==3.4.0